+++ packages/components/src/form/InputWrapper/InputWrapper.constants.js
@@ -22,6 +22,7 @@ export const INPUT_WRAPPER_PROP_TYPES = {
   contentStyle: PropTypes.any,
   autoComplete: PropTypes.string,
   showEmptyLabel: PropTypes.bool,
+  readOnly: PropTypes.bool,
 };
 export const INPUT_WRAPPER_DEFAULT_PROPS = {
   label: '',
@@ -33,4 +34,5 @@ export const INPUT_WRAPPER_DEFAULT_PROPS = {
   required: false,
   autoComplete: 'off',
   disabled: false,
+  readOnly: false,
 };