+++ packages/components/src/layout/TotalLayout/TotalLayoutStepContainer/TotalLayoutStepContainer.styles.js
@@ -2,10 +2,10 @@ import { createStyles } from '@mantine/styles';
 import { TOTAL_LAYOUT_HEADER_HEIGHT } from '../TotalLayoutHeader/TotalLayoutHeader.constants';
 
 const TotalLayoutStepContainerStyles = createStyles(
-  (theme, { hasFooter, clean, fullWidth, noMargin, footerPadding }) => ({
+  (theme, { hasFooter, clean, fullWidth, noMargin, footerPadding, TopZone, forceNotMaxWidth }) => ({
     root: {},
     stepContainer: {
-      padding: clean ? 0 : '30px 0 0 0 ',
+      padding: clean || TopZone ? 0 : '30px 0 0 0 ',
       marginLeft: fullWidth && !noMargin ? theme.spacing[8] : 0,
       marginRight: fullWidth && !noMargin ? theme.spacing[8] : 0,
       height: '100%',
@@ -21,6 +21,10 @@ const TotalLayoutStepContainerStyles = createStyles(
     },
     stepName: {
       marginBottom: 12,
+    },
+    topZone: {
+      marginBottom: 12,
+      marginTop: 24,
     },
   }),
 );