+++ packages/components/src/overlay/LoadingOverlay/LoadingOverlay.styles.js
@@ -3,7 +3,7 @@ import { createStyles } from '@mantine/styles';
 const LoadingOverlayStyles = createStyles((theme) => ({
   root: {
     svg: {
-      stroke: theme.other.core.color.primary['300'],
+      stroke: theme.other.core?.color?.primary['300'],
     },
   },
 }));