+++ packages/editors/src/form/TextEditorInput/TextEditorInput.constants.js
@@ -14,6 +14,17 @@ export const TEXT_EDITOR_DEFAULT_TOOLBARS = {
 export const TEXT_EDITOR_TEXTAREA_TOOLBARS = {
   style: true,
   heading: false,
+  align: false,
+  list: false,
+  history: false,
+  color: false,
+  formulation: false,
+  link: true,
+};
+
+export const TEXT_EDITOR_NO_HEADINGS_TOOLBARS = {
+  style: true,
+  heading: false,
   align: true,
   list: true,
   history: true,